Remove blank-user-message util and inline repairs
Delete the standalone blank-user-message module and its tests, inline the blank-user-message repair logic into session-file-repair, and remove in-memory blank-message filtering/warning from the replay/sanitization path. Updates: removed src/agents/blank-user-message.ts and its test; adjusted sanitize-session-history test harness to drop the optional warn callback; removed tests that asserted warn was called; removed use of dropBlankUserMessages from replay-history and replay attempt code; inlined content-repair logic into session-file-repair.ts to keep on-disk repair behavior. These changes centralize blank-message handling to file-repair and eliminate the prior in-memory pre-filter and warning.
