+++ src/agents/pi-embedded-runner.sanitize-session-history.test-harness.ts
@@ -13,7 +13,6 @@ export type SanitizeSessionHistoryFn = (params: {
   sessionId: string;
   modelId?: string;
   policy?: TranscriptPolicy;
-  warn?: (message: string) => void;
 }) => Promise<AgentMessage[]>;
 export type SanitizeSessionHistoryMockedHelpers = typeof import("./pi-embedded-helpers.js");
 export type SanitizeSessionHistoryHarness = {
