+++ src/agents/pi-embedded-runner/run/attempt.ts
@@ -1919,7 +1919,6 @@ export async function runEmbeddedAttempt(
             sessionManager,
             sessionId: params.sessionId,
             policy: transcriptPolicy,
-            warn: (message) => log.warn(message),
           });
           cacheTrace?.recordStage("session:sanitized", { messages: prior });
           const validated = await validateReplayTurns({
